clean code
# message auto-generated for no-merge-commit merge:
merge cleancode0609 into master

cleancode清理

Created-by: insightDeveloper

Match-id-774ca17f30bba1b276510c849a9c6aa2be368794

+++ server/src/modules/summary/protocol/SummaryProtocolResponse.h
@@ -192,7 +192,7 @@ struct QueryParallelStrategyResponse : public Response {
     QueryParallelStrategyResponse() : Response(REQ_RES_SUMMARY_QUERY_PARALLEL_STRATEGY) {}
     Module::ParallelStrategyConfig config;
     std::string level;
-    const int64_t validValue = 1;
+    const uint32_t validValue = 1;
     bool IsValid() const
     {
         if (config.ppSize < validValue || config.tpSize < validValue || config.dpSize < validValue) {
